+++ jest.config.js
@@ -17,8 +17,8 @@ module.exports = {
   },
   coverageThreshold: {
     global: {
-      branches: 50,
-      functions: 80,
+      branches: 60,
+      functions: 75,
       lines: 80,
       statements: 80,
     },